Move the dependency walk into core: resolveInputs + source-resolver ports
Stage 4a of the unified-runtime migration. The ~320-line body of
resolveDependencies moves from index.ts into the corpus-agnostic
resolveInputs in @corpus/core/run/producer-run.ts, ported line-by-line:
same Promise.all parallelism, cycle-detection point and exact message,
error-value shapes, fanOut per-instance semantics, sourcesOnly
transitive-closure recursion, recordSource + the 8000-char preview cap.

App-specific behavior enters through ports: the 8 string-dependency
branches become source-resolver closures in the new
src/worker/run-sources.ts (gemara, commentaries, mishna with
per-instance filtering, halacha-refs, yerushalmi-text, incoming,
context/context-light sharing one closure), and the recursive
mark/enrichment runs are injected callbacks (runMarkOnce /
runEnrichmentOnce stay in index.ts for stage 4b to unify).

resolveDependencies keeps its exported signature as a thin delegate;
index.ts shrinks by 418 lines. Zero test edits, zero snapshot churn —
the PR #356 characterization suite passes byte-identically.
